Add median to stdlib [KT-50138]
Although figuring out the median is trivial when the numbers are sorted,
it is not that straight-forward to do it in a performant way when the
numbers are not guaranteed to be sorted.

The naive approach would be to just sort it and then get the median
value. However, sorting is suboptimal, with a time complexity of
O(n.log(n)).

For calculating the median, QuickSelect can be used to perform the task
in linear time: O(n).

By adding this to the stdlib, developers can rely on a standard
implementation instead of having to come up with an implementation of
their own.

^KT-50138 Fixed