Generate stability field for known stable classes
This fixes an issue where marking a class as `@Stable` can cause an ABI
incompatibility. If a class was previously inferred as being runtime
stable, we relied on the static `$stability` property to tell its
stability. If the class then became stable, the static property would no
longer be generated. For pre-compiled libraries, this can result in
unresolved references to this field.

With this CL, we'll now start to generate the `$stability` field for all
classes annotated with `@Stable`. This field should be stripped with R8
in release builds of applications, so the added generation shouldn't add
much additional overhead while ensuring ABI compatibility when a class
becomes `@Stable`.

Relnote: """Fixed an issue where annotating a class as `@Stable` would
    not be an ABI-compatible change.
    """

Fixes: 264552542 ( https://issuetracker.google.com/issues/264552542 )
Test: ./gradlew :compose:compiler:c-h:i-t:tDUT
Change-Id: I30172b6d208d65dd2213f81b746a2f8c4836432a ( https://android-review.git.corp.google.com/q/I30172b6d208d65dd2213f81b746a2f8c4836432a )