[Gradle] Update test that resolves jsMain dependencies metadata
This behavior is incorrect and it is not supported by new
resolvable metadata dependencies configurations.
Because this configuration is expected to be resolved to metadata
artifacts only. And the fact that it gets resolvled to some platform
variant is not right. But is not critical as well.

The next steps is to remove metadata configurations from platform
source sets.