Add primitive versions of State and MutableState
This CL adds APIs for (Mutable)IntState, (Mutable)LongState,
(Mutable)FloatState, and (Mutable)DoubleState. The goal here is to offer
variations of our existing (`Mutable`)`State<T>` APIs that do not
experience the penalty of autoboxing when used to store primitive
values.

We are not adding `BooleanState` here because the autoboxing penalty for
`Boolean` types is virtually nonexistent because the JVM manages a pool
of the boxed versions of true and false, meaning that the boxing
operation will never cause an allocation.

Relnote: """Added primitive versions of the `State` API, allowing Int,
    Long, Float, and Double values to be tracked in `State` objects
    without incurring penalties for autoboxing.
    """

Test: ./gradlew :compose:runtime:r:tDUT
Change-Id: I48e438d0af15226217b77b2ba9a970c3219580bf ( https://android-review.git.corp.google.com/q/I48e438d0af15226217b77b2ba9a970c3219580bf )