Remove skipping code generation from inline lambda
Remove the code to skip the body of inline lambdas. This
code was incorrect as it could cause code to be skipped that
was intneded to be executed because it was invalidated by
a state read. As inline lambdas cannot be restarted they
also cannot be skipped.

Fixes: 264710012 ( https://issuetracker.google.com/issues/264710012 )
Test: ./gradlew :compsoe:compiler:c-h:i-t:tDUT

Change-Id: I2c44ea4a15a04193f535c084c9af85b0c36da04a ( https://android-review.googlesource.com/q/I2c44ea4a15a04193f535c084c9af85b0c36da04a )

Moved from: https://github.com/androidx/androidx/commit/9c7c15679b050abb35dd0d7a786ce617317236c2