Add defensive fallback if checking Kotlin compatibility fails.
Kotlin 1.8.0 seems to have changed the way `MessageCollector.report` is resolved, which caused us to fail exceptionally when attempting to check for a version mismatch.  Building on top of shifting sands, there is not much we can do to reasonably defend against any given function randomly not working.  In such cases, it seems likely that it's a version check incompatibility, so we defensively wrap the entire version checker in a try-catch block in order to hint at the possible problem/solution and rethrow the original trace.

Change-Id: Ie68f7790b31dc40ef181fc72a5c4038a3beae3d8 ( https://android-review.git.corp.google.com/q/Ie68f7790b31dc40ef181fc72a5c4038a3beae3d8 )