Add capability to fetch coroutine stack trace info in one call
Previously, when the Kotlin debugger needed to fetch a coroutine async stack trace, it needed to perform a lot of JDI calls to unwind a call stack and collect information about each stack frame individually. That led to very slow stack trace loading time when debugging an Android app on a phone, since the JDI communication with a device is very expensive. This commit adds a method that fetches all of the required information about a stack trace on the client VM side, which significantly speeds up the stack trace loading time.