JVM_IR: Fix suspend function tail-call optimization glitch
Since the return type of `fn` is not guaranteed to be Unit at runtime
in such case, we can't resolve the issue by popping out the return
value and pushing Unit onto the stack.

Thus, we need to disable the tail-call optimization, and generate a
state machine for suspend function `bar`.