[FIR] KT-56725: Don't add a named arg to the set-operator
Transforming the indexed set last argument into a
named argument is very-very confusing.
Also, the argument mapping code isn't that
obvious, so let's have a one single confusing
place instead of two.

When we jump from the current parameter to another one, we can't put the
`set` value into the result and not
change the `currentPositionedParameterIndex`,
expecting that it will be closed later anyway.
This is because the order of putting things into `result`
matters, so we must first close the vararg.

^KT-56725 Fixed