Add test on correct module metadata produced with K2
It's not a very good test because it just checks what is produced by the
compiler, without the real effect that it has when a compiler,
potentially of other version, will read it. But the latter is difficult
to achieve because there's no infrastructure in compiler tests for
running any other compiler version than the current.

 #KT-56526