[FIR] Implement the lazy supertype traverser
Replace the notion of a representative superclass
with the least upper bound.
This makes complex types like
intersection/flexible transparent to
RULES1-based compatibility checks.
This also means that a `ConeKotlinType` may be both
`isClass` and `isEnumClass` (when it's an intersection type).
One way to look at it is to think
that this is an automatic way of handling
type parameters: automatic picking of
"interesting" bounds, and checking them against one another.