Compose: Move K1 specific code to a separate package
Ideally, there should be no references from the Compose IR plugin to
K1 specific code. Currently, there are still two references from
`ComposableSymbolRemapper` and `CopyDefaultValuesFromExpectLowering`,
but these passes are K1 specific anyway and shouldn't run on K2.

The remaining `BindingContext` usages in `ComposeWritableSlices` don't
refer to any frontend data structures and are only used internally in
the Compose IR plugin. Technically, this is still a K1 specific API and
we should remove it, but it's harmless for now.

Bug: 271370276 ( https://issuetracker.google.com/issues/271370276 )
Test: ./gradlew :compose:compiler:c-h:i-t:tDUT
Change-Id: I553d66a8ae74cf4b32b6faec95d4a88da65e00a2 ( https://android-review.git.corp.google.com/q/I553d66a8ae74cf4b32b6faec95d4a88da65e00a2 )