fixup! [FIR] KT-57104: Preserve @JvmName on fake-override accessors
`FirPsiIrTextTestGenerated.testDelegatedGenericImplementation`
used to render type parameters from `<get-x>`,
because when assigning `extensionReceiverParameter`
of the setter `<set-x>` we came to
`Fir2IrClassifierStorage.kt:638`,
and in this cache there's the parameter with
the `<get-x>` parent.
Note that `typeContext.origin == DEFAULT` in
`getCachedIrTypeParameter`.
It's `DEFAULT` because at the line `Fir2IrDeclarationStorage.kt:335`
the `function` variable was `null` (because there's no setter).